+++ framework-core/src/main/java/io/summerx/framework/cache/concurrent/ConcurrentMapCustomizedCache.java
@@ -32,12 +32,12 @@ public class ConcurrentMapCustomizedCache extends ConcurrentMapCache implements 
     }
 
     @Override
-    public void put(Object key, Object value, long expired) {
+    public void put(Object key, Object value, int expired) {
         put(key, value);
     }
 
     @Override
-    public void putIfAbsent(Object key, Object value, long expired) {
+    public void putIfAbsent(Object key, Object value, int expired) {
         putIfAbsent(key, value);
     }
